llvmpipe: use 32-bit size for stride instead of 64-bit

+++ src/gallium/drivers/llvmpipe/lp_texture.h
@@ -83,7 +83,7 @@ struct llvmpipe_resource
    /**
     * Stride of each sample
     */
-   uint64_t sample_stride;
+   unsigned sample_stride;
 
    /**
     * Malloc'ed data for regular textures, or a mapping to dt above.
